==> health services/2 sti models - all health services.Rmd <==
---
title: "Modelos de serie de tiempo interrumpida para caída en prestaciones oncológicas"
output: 
---

#

```{r LIBRERIAS, message=FALSE, warning=FALSE, include=FALSE}
library(data.table)
library(tidyverse)
library(MASS)
library(sjPlot)
```

# 1. Datos

```{r LOAD AND PREPARE SAVED DATA GROUPED BY WEEK}
load("output/DB/DB para modelos agregada.RData")

prest <- levels(dat.ag[,grupo.trz])
prest.sexo <- as.character(unique(dat.ag[, .(grupo.trz = grupo.trz, grp.cancer = grp.cancer)][
                            !(grp.cancer %in% c("Prostate Cancer", "Breast Cancer", 
                                                "Cervical Cancer"))][
                                       order(grupo.trz)  ][ , grupo.trz]))
dat.ag <- dat.ag[
  , week := as.factor(week)][
  !(grupo.trz == prest[5] & edad_cat == "[0,10)")][
  !(grupo.trz == prest[8] & edad_cat == "[0,10)")][
  !(grupo.trz == prest[15] & edad_cat %in% c("[0,10)", "[10,20)"))][
  , pandemia := cuarentenas][, cuarentenas := NULL][
  , prevision := factor(prevision, labels = c("Public", "Private"))]

```


# 2. Modelo efectos fijos

```{r MODELO AGREGADO ADITIVO Y POR SUBSETS, message=F}

# Agregado y por subsets

sets <- list(
  dat.ag,
  dat.ag[grupo.trz %in% prest.sexo][sexo == "Female"],
  dat.ag[grupo.trz %in% prest.sexo][sexo == "Male"],
  dat.ag[prevision == "Private"],
  dat.ag[prevision == "Public"],
  dat.ag[edad_cat == "[0,10)"],
  dat.ag[edad_cat == "[10,20)"],
  dat.ag[edad_cat == "[20,30)"],
  dat.ag[edad_cat == "[30,40)"],
  dat.ag[edad_cat == "[40,50)"],
  dat.ag[edad_cat == "[50,60)"],
  dat.ag[edad_cat == "[60,70)"],
  dat.ag[edad_cat == "[70,80)"]  )

formulas <- list(
  as.formula(count ~ sexo + edad_cat + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ edad_cat + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ edad_cat + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + edad_cat + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + edad_cat + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after),
  as.formula(count ~ sexo + prevision + year + week + pandemia + time + pandemia:time_after)
)

models <- list()

for (i in 1:13){
  
  sets[[i]] <- sets[[i]] %>% 
    group_by(date_wk, sexo, edad_cat, prevision, year, week, time, pandemia, time_after) %>% 
    summarise(count = sum(count))

  models[[i]] <- glm.nb(formulas[[i]], data = sets[[i]])
  
}

save(models, file = "output/modelos/sti agregados.RData" )

```

## 2.1 Tablas modelos

```{r}
load(file = "output/modelos/sti agregados.RData")

tab.models <- data.table()

for (i in 1:13){

  summ <- summary(models[[i]])
  mod.res <- as.data.frame(summ$coefficients)
  mod.res$Model <- i
  mod.res$Vars <- rownames(summ$coefficients)

  tab.models <- rbind(tab.models, mod.res)
  
}

rm(summ, mod.res, i)

confint.models <- data.table()

for (i in 1:13){
  
  conf <- confint(models[[i]])
  conf.res <- as.data.frame(conf)
  conf.res$Model <- i
  conf.res$Vars <- rownames(conf)

  confint.models <- rbind(confint.models, conf.res)
  
}

rm(conf, conf.res, i)

tab.models <- cbind(tab.models, confint.models[,c(1:2)])

invlink <- models[[1]]$family$linkinv

. <- tab.models[
  , IRR := invlink(Estimate) ][
  , IC_lower := invlink(`2.5 %`) ][
  , IC_upper := invlink(`97.5 %`) ]

tab.models <- .

save(tab.models, file = "output/tabs/coef sti datos agregados.RData")

```

## 2.2 Gráfico de coeficientes de modelos

```{r fig.height=10, fig.width=6}
load(file = "output/tabs/coef sti datos agregados.RData")

tab.models %>% 
  ggplot(aes(x = IRR, y = Vars)) +
  geom_point()
```

```{r}
tab.plot <- tab.models %>% 
  filter(Vars %in% c("(Intercept)", "edad_cat[0,10)", "edad_cat[10,20)", "edad_cat[20,30)", 
                       "edad_cat[30,40)", "edad_cat[50,60)", "edad_cat[60,70)", "edad_cat[70,80)", 
                       "pandemia", "pandemia:time_after", "previsionPrivate", "sexoFemale", "time")) %>% 
  mutate(Mod.int = as.factor(Model),
         Mod.int = factor(Mod.int, labels = c("Additive", rep("Sex", 2), rep("Insurance", 2), rep("Age", 8))),
         Model = as.factor(Model),
         Model = factor(Model, labels = c("Additive", "Female", "Male", "Private", "Public", "[0,10)", "[10,20)",
                                          "[20,30)", "[30,40)", "[40,50)", "[50,60)", "[60,70)", "[70,80)"))) %>% 
  ggplot(aes(x = IRR, y = Vars)) +
  geom_point(aes(color = Model)) +
  geom_errorbarh(aes(xmin = IC_lower, xmax = IC_upper, color = Model), alpha = .25) +
  geom_vline(xintercept = 1, alpha = .3, linetype = 2) +
  facet_grid(. ~ Mod.int, scales = "free_x")

tab.plot

ggsave(plot = tab.plot,
       filename = "output/plots/tab plot modelos.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)
```


# 3. Predicciones

```{r PREDICTED DB}
dat.plots.list <- list()

for (i in 1:13) {
  
  dat.cf <- sets[[i]]
  dat.cf$pandemia <- 0

  dat.plots.temp <- sets[[i]]

  ilink <- family(models[[i]])$linkinv
  
  pred <- predict(models[[i]], se.fit = T)
  yhat <- pred$fit
  se <- pred$se.fit
  
  dat.plots.temp$yhat <- ilink(yhat)
  dat.plots.temp$ub <- ilink(yhat + (1.96*se))
  dat.plots.temp$lb <- ilink(yhat - (1.96*se))
  
  
  pred.cf <- predict(models[[i]], newdata = dat.cf, se.fit = T)
  cf <- pred.cf$fit
  cf.se <- pred.cf$se.fit
  
  dat.plots.temp$cf <- ilink(cf)
  dat.plots.temp$cf_ub <- ilink(cf + (1.96*cf.se))
  dat.plots.temp$cf_lb <- ilink(cf - (1.96*cf.se))
  
  dat.plots.list[[i]] <- dat.plots.temp
  
  rm(dat.cf, dat.plots.temp, ilink, pred, yhat, se, pred.cf, cf, cf.se)
}

```

```{r RECODE AND BIND PREDICTED DB}
mod <- c("Additive", 
         rep("Sex", 2), 
         rep("Insurance", 2), 
         rep("Age", 8))

for (i in 1:13) {

  dat.plots.list[[i]]$mod.int <- mod[i]
  
}

dat.plots <- data.table()

for (i in 1:13){
  
  dat.plots <- rbind(dat.plots, dat.plots.list[[i]])
  
}

rm(dat.plots.list)

save(dat.plots, file = "output/DB/DB predichos nb sti.RData")

```

# 4. Plots

```{r}
load("output/DB/DB predichos nb sti.RData")
```


## 4.1 Serie completa

```{r}
g1.a <- dat.plots %>% 
  filter(mod.int == "Additive") %>% 
  group_by(date_wk) %>% summarise(count = sum(count),
                                  yhat = sum(yhat),
                                  cf = sum(cf),
                                  ub = sum(ub),
                                  lb = sum(lb),
                                  cf_ub = sum(cf_ub),
                                  cf_lb = sum(cf_lb)) %>% 
  ggplot(aes(x = date_wk, y = count)) +
  geom_point(alpha = .3) +
  geom_line(aes(x = date_wk, y = yhat)) +
  geom_ribbon(aes(ymin = lb, ymax = ub), alpha = .3) +
  geom_line(aes(x = date_wk, y = cf), linetype = 2, size = .3) +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub), alpha = .15) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Date", 
       title = "Additive model") +
  geom_vline(xintercept = as.Date("2020-03-15"), linetype = "dotted", color = "red") +
  theme(legend.position = "bottom",
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 9),
        plot.subtitle = element_text(size = 14),
        plot.title = element_text(size = 12),
        strip.text = element_text(size = 14))
  
g1.a
```

```{r}
g1.b <- dat.plots %>% 
  filter(mod.int == "Sex") %>%
  mutate(sexo = fct_relevel(sexo, sort)) %>% 
  group_by(date_wk, sexo) %>% summarise(count = sum(count),
                                  yhat = sum(yhat),
                                  cf = sum(cf),
                                  ub = sum(ub),
                                  lb = sum(lb),
                                  cf_ub = sum(cf_ub),
                                  cf_lb = sum(cf_lb)) %>% 
  ggplot(aes(x = date_wk, y = count)) +
  geom_point(alpha = .3, aes(color = sexo)) +
  geom_line(aes(x = date_wk, y = yhat, color = sexo)) +
  geom_ribbon(aes(ymin = lb, ymax = ub, fill = sexo), alpha = .3) +
  geom_line(aes(x = date_wk, y = cf, color = sexo), linetype = 2, size = .3) +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, fill = sexo), alpha = .15) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Date", fill = "Sexo", color = "Sexo",
       title = "by sex") +
  geom_vline(xintercept = as.Date("2020-03-15"), linetype = "dotted", color = "red") +
  theme(legend.position = "bottom",
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 9),
        plot.subtitle = element_text(size = 14),
        plot.title = element_text(size = 12),
        strip.text = element_text(size = 14))
  
g1.b
```

```{r}
g1.c <- dat.plots %>% 
  filter(mod.int == "Insurance") %>% 
  group_by(date_wk, prevision) %>% summarise(count = sum(count),
                                  yhat = sum(yhat),
                                  cf = sum(cf),
                                  ub = sum(ub),
                                  lb = sum(lb),
                                  cf_ub = sum(cf_ub),
                                  cf_lb = sum(cf_lb)) %>% 
  ggplot(aes(x = date_wk, y = count)) +
  geom_point(aes(color = prevision), alpha = .3) +
  geom_line(aes(x = date_wk, y = yhat, color = prevision)) +
  geom_ribbon(aes(ymin = lb, ymax = ub, fill = prevision), alpha = .3) +
  geom_line(aes(x = date_wk, y = cf, color = prevision), linetype = 2, size = .3) +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, fill = prevision), alpha = .15) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Date", fill = "Previsión", color = "Previsión",
       title = "by insurance") +
  geom_vline(xintercept = as.Date("2020-03-15"), linetype = "dotted", color = "red") +
  theme(legend.position = "bottom",
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 9),
        plot.subtitle = element_text(size = 14),
        plot.title = element_text(size = 12),
        strip.text = element_text(size = 14))
  
g1.c
```

```{r}
g1.d <- dat.plots %>% 
  filter(mod.int == "Age") %>% 
  mutate(edad_cat = fct_relevel(edad_cat, sort) ) %>% 
  group_by(date_wk, edad_cat) %>% summarise(count = sum(count),
                                  yhat = sum(yhat),
                                  cf = sum(cf),
                                  ub = sum(ub),
                                  lb = sum(lb),
                                  cf_ub = sum(cf_ub),
                                  cf_lb = sum(cf_lb)) %>% 
  ggplot(aes(x = date_wk, y = count)) +
  geom_point(aes(color = edad_cat), alpha = .15) +
  geom_line(aes(x = date_wk, y = yhat, color = edad_cat)) +
  geom_ribbon(aes(ymin = lb, ymax = ub, fill = edad_cat), alpha = .3) +
  geom_line(aes(x = date_wk, y = cf, color = edad_cat), linetype = 2, size = .3) +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, fill = edad_cat), alpha = .075) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Date", fill = "Edad", color = "Edad",
       title = "by age") +
  geom_vline(xintercept = as.Date("2020-03-15"), linetype = "dotted", color = "red") +
  theme(legend.position = "bottom",
        axis.title = element_text(size = 10),
        axis.text = element_text(size = 9),
        plot.subtitle = element_text(size = 14),
        plot.title = element_text(size = 12),
        strip.text = element_text(size = 14))
  
g1.d
```

## 4.2 Exposición

```{r}
g2.a <- dat.plots %>% 
  mutate(week = as.numeric(week)) %>% 
  filter(mod.int == "Additive") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019"),
         cf = ifelse(exposicion == "2020" & pandemia != 0, cf, NA),
         cf_ub = ifelse(exposicion == "2020" & pandemia != 0, cf_ub, NA),
         cf_lb = ifelse(exposicion == "2020" & pandemia != 0, cf_lb, NA) ) %>% 
  group_by(week, year, exposicion) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = sum(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  
  ggplot(aes(x = week, y = count, fill = exposicion))+
  
  geom_point(alpha = .3, aes(color = exposicion), size = 1.5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +

  geom_line(aes(x = week, y = cf, color = exposicion, linetype = "Predicted counterfactual"), 
            size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       title = "Health Services associated with Cancer") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 10),
        plot.title = element_text(size = 12))

g2.a
```

```{r}
g2.b <- dat.plots %>% 
  mutate(week = as.numeric(week)) %>% 
  filter(mod.int == "Sex") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019"),
         cf = ifelse(exposicion == "2020" & pandemia != 0, cf, NA),
         cf_ub = ifelse(exposicion == "2020" & pandemia != 0, cf_ub, NA),
         cf_lb = ifelse(exposicion == "2020" & pandemia != 0, cf_lb, NA)) %>% 
  group_by(week, year, exposicion, sexo) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, sexo)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = sum(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  
  geom_point(alpha = .3, aes(color = exposicion), size = 1.5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +

  geom_line(aes(x = week, y = cf, color = exposicion, linetype = "Predicted counterfactual"), size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  
  facet_grid(. ~ sexo) + 
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       title = "Health Services associated with Cancer",
       subtitle = "by sex") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 10),
        plot.title = element_text(size = 12))

g2.b
```

```{r}
g2.c <- dat.plots %>% 
  mutate(week = as.numeric(week)) %>% 
  filter(mod.int == "Insurance") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019"),
         cf = ifelse(exposicion == "2020" & pandemia != 0, cf, NA),
         cf_ub = ifelse(exposicion == "2020" & pandemia != 0, cf_ub, NA),
         cf_lb = ifelse(exposicion == "2020" & pandemia != 0, cf_lb, NA)) %>% 
  group_by(week, year, exposicion, prevision) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, prevision)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = sum(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  
  geom_point(alpha = .3, aes(color = exposicion), size = 1.5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +

  geom_line(aes(x = week, y = cf, color = exposicion, linetype = "Predicted counterfactual"), size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  
  facet_grid(. ~ prevision) + 
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       title = "Health Services associated with Cancer",
       subtitle = "by insurance") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 10),
        plot.title = element_text(size = 12))

g2.c
```

```{r}
g2.d <- dat.plots %>% 
  mutate(week = as.numeric(week)) %>% 
  filter(mod.int == "Age") %>% 
  mutate(edad_cat = fct_relevel(edad_cat, sort),
         exposicion = ifelse(year == 2020, "2020", "2018-2019"),
         cf = ifelse(exposicion == "2020" & pandemia != 0, cf, NA),
         cf_ub = ifelse(exposicion == "2020" & pandemia != 0, cf_ub, NA),
         cf_lb = ifelse(exposicion == "2020" & pandemia != 0, cf_lb, NA)) %>% 
  group_by(week, year, exposicion, edad_cat) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, edad_cat)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = sum(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  
  geom_point(alpha = .3, aes(color = exposicion), size = 1.5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +

  geom_line(aes(x = week, y = cf, color = exposicion, linetype = "Predicted counterfactual"), size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  
  facet_wrap("edad_cat", nrow = 2) + 
  
  ylim(0, NA) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       title = "Health Services associated with Cancer",
       subtitle = "by insurance") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 10),
        plot.title = element_text(size = 12))

g2.d
```

Guardar gráficos
```{r}
ggsave(plot = g1.a,
       filename = "output/plots/serie completa aditivo.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

ggsave(plot = g1.b,
       filename = "output/plots/serie completa sexo.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

ggsave(plot = g1.c,
       filename = "output/plots/serie completa prevision.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

ggsave(plot = g1.d,
       filename = "output/plots/serie completa edad.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

g1 <- ggpubr::ggarrange(g1.a, g1.b, g1.c, g1.d, ncol = 2, nrow = 2)

ggsave(plot = g1,
       filename = "output/plots/serie completa.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)
```

```{r}
ggsave(plot = g2.a,
       filename = "output/plots/serie 2020 vs anterior aditivo.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

ggsave(plot = g2.b,
       filename = "output/plots/serie 2020 vs anterior sexo.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

ggsave(plot = g2.c,
       filename = "output/plots/serie 2020 vs anterior prevision.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

ggsave(plot = g2.d,
       filename = "output/plots/serie 2020 vs anterior edad.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)

g2 <- ggpubr::ggarrange(g2.a, g2.b, g2.c, g2.d, ncol = 2, nrow = 2, common.legend = T)

ggsave(plot = g2,
       filename = "output/plots/serie 2020 vs anterior.png", 
       device = "png",
       dpi = "retina",
       width = 12, height = 8)
```
