==> health services/7 sti plots.Rmd <==
---
title: "Figuras de resultados modelos de serie de tiempo interrumpida para caída en prestaciones oncológicas"
subtitle: "Binomial negativo agregado y por tipo de prestación; submodelos por sexo, previsión y edad"
output: 
---


#

```{r LIBRERIAS, message=FALSE, warning=FALSE, include=FALSE}
rm(list = ls())
library(data.table)
library(tidyverse)
library(ggpubr)
```

# Datos

```{r DATA LABELS}
prest <- c("Oncology out-patient visits", "Cancer-related antigen test", "Biopsies", "Neoplastic Cells test" , "PET-CT", 
            "Fecal occult bleeding test", "Sigmoidoscopy and\nColonoscopy", 
            "Cervical Biopsy", "Colposcopy", "Papanicolaou", 
            "Esophagoscopy,\nGastroduodenoscopy,\nYeyuno Ileoscopy", "Ureasa tests", 
            "Breast Ultrasound", "Mammography", "Breast MRI",
            "Prostate-Specific\nAntigen", "Prostate Ultrasound")

prest.sexo <- c("Oncology out-patient visits", "Cancer-related antigen test", "Biopsies", "Neoplastic Cells test" , "PET-CT", 
            "Fecal occult bleeding test", "Sigmoidoscopy and\nColonoscopy", 
            "Esophagoscopy, Gastroduodenoscopy, Yeyuno Ileoscopy", "Ureasa tests")
```

```{r LOAD DATA PLOTS}
load("output/DB/DB predichos nb sti.RData")
dat.plots.agg <- dat.plots

load(file = "output/DB/DB predichos nb sti x prestacion.RData")
dat.plots.type <- dat.plots[
  , grupo.trz := factor(grupo.trz, labels = prest) ]


load(file = "output/DB/DB predichos nb sti x sexo.RData")
dat.plots.type.sexo <- dat.plots[
  , grupo.trz := factor(grupo.trz, labels = prest.sexo) ]

load(file = "output/DB/DB predichos nb sti x prevision.RData")
dat.plots.type.prev <- dat.plots[
  , grupo.trz := factor(grupo.trz, labels = prest) ]

rm(dat.plots)

```


# Figuras

## Figura 1

```{r FIG1, fig.width=6, fig.height=9, message=FALSE, warning=FALSE}
g1.agg <- dat.plots.agg %>% 
  filter(mod.int == "Additive") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019")) %>% 
  group_by(week, year, exposicion) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = mean(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb) ) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  geom_point(alpha = .3, aes(color = exposicion), size = 1) +
  geom_line(aes(x = week, y = yhat, color = exposicion, group = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +
  geom_line(aes(x = week, y = cf, color = exposicion, group = exposicion, linetype = "Predicted counterfactual"), 
            size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  ylim(0, NA) +
  scale_x_discrete(breaks = seq(5, 53, 6)) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       subtitle = "", title = "Total cancer-related outpatient health services") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 7),
        plot.title = element_text(size = 11, face = "bold"))

plots.g1 <- list()

for (i in 1:17){
  plots.g1[[i]] <- dat.plots.type %>% 
  filter(grupo.trz == levels(dat.plots.type$grupo.trz)[i]) %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019")) %>% 
  group_by(week, year, exposicion, grupo.trz) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, grupo.trz) %>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = mean(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb)) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  geom_point(alpha = .3, aes(color = exposicion), size = .5) +
  geom_line(aes(x = week, y = yhat, color = exposicion, group = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +
  geom_line(aes(x = week, y = cf, color = exposicion, group = exposicion, linetype = "Predicted counterfactual"), 
            size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  ylim(0, NA) +
  scale_x_discrete(breaks = seq(11, 53, 12)) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       subtitle = levels(dat.plots.type$grupo.trz)[i], title = "") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 7),
        plot.title = element_text(size = 8, face = "bold"))
}


g1.a <- ggarrange(plots.g1[[1]] + ggtitle("Non-specific services"), 
                  plots.g1[[2]], plots.g1[[3]], plots.g1[[5]], 
                  ncol = 4, nrow = 1, common.legend = T, legend = "none", align = "hv")
g1.b <- ggarrange(plots.g1[[6]] + ggtitle(levels(dat.plots.type$grp.cancer)[2]), plots.g1[[7]], 
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")
g1.c <- ggarrange(plots.g1[[12]] + ggtitle("Stomach cancer"), plots.g1[[11]], 
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")
g1.d <- ggarrange(plots.g1[[8]] + ggtitle(levels(dat.plots.type$grp.cancer)[3]), plots.g1[[9]], plots.g1[[10]], 
                  ncol = 4, nrow = 1, common.legend = T, legend = "none", align = "hv")
g1.e <- ggarrange(plots.g1[[13]] + ggtitle(levels(dat.plots.type$grp.cancer)[5]), plots.g1[[14]],  
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")
g1.f <- ggarrange(plots.g1[[17]] + ggtitle(levels(dat.plots.type$grp.cancer)[6]), plots.g1[[16]], 
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")

g1.bc <- ggarrange(g1.b, g1.c)
g1.ef <- ggarrange(g1.e, g1.f)

g1.type <- ggarrange(g1.a, g1.bc, g1.ef, g1.d, ncol = 1, 
          common.legend = T, legend = "top")

g1 <- ggarrange(g1.agg, g1.type, ncol = 1, heights = c(1, 3),
          common.legend = T, legend.grob = get_legend(plots.g1[[1]]), legend = "top")
g1

ggsave(plot = g1,
       filename = "output/plots/paper prestaciones fig1.pdf", 
       device = "pdf",
       dpi = "retina",
       width = 8, height = 12)

rm(g1.a, g1.b, g1.c, g1.d, g1.e, g1.f, g1.bc, g1.ef, g1.type, g1.agg)

g1
```


# Figuras suplementarias

## Figura 1 suplementaria por sexo y previsión

```{r FIG1 POR PREVISION, fig.width=6, fig.height=18, message=FALSE, warning=FALSE}
g1.agg <- dat.plots.agg %>% 
  filter(mod.int == "Insurance") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019")) %>% 
  group_by(week, year, exposicion, prevision) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, prevision)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = mean(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb)) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  geom_point(alpha = .3, aes(color = exposicion), size = .7) +
  geom_line(aes(x = week, y = yhat, color = exposicion, group = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +
  geom_line(aes(x = week, y = cf, color = exposicion, group = exposicion, linetype = "Predicted counterfactual"), 
            size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  ylim(0, NA) +
  scale_x_discrete(breaks = seq(5, 53, 6)) +
  facet_grid(. ~ prevision) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       subtitle = "", title = "Total cancer-related health services") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 8),
        plot.title = element_text(size = 11, face = "bold"))

plots.g1 <- list()

for (i in 1:17){
  plots.g1[[i]] <- dat.plots.type.prev %>% 
  filter(grupo.trz == levels(dat.plots.type.prev$grupo.trz)[i]) %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019")) %>% 
  group_by(week, year, exposicion, grupo.trz, prevision) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, grupo.trz, prevision) %>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = mean(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb)) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  #geom_point(alpha = .3, aes(color = exposicion), size = .7) +
  geom_line(aes(x = week, y = yhat, color = exposicion, group = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +
  geom_line(aes(x = week, y = cf, color = exposicion, group = exposicion, linetype = "Predicted counterfactual"), 
            size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  ylim(0, NA) +
  scale_x_discrete(breaks = seq(11, 53, 12)) +
  facet_grid(. ~ prevision) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       subtitle = levels(dat.plots.type.prev$grupo.trz)[i], title = "") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 8),
        plot.title = element_text(size = 11, face = "bold"))
}


g1.a <- ggarrange(plots.g1[[1]] + ggtitle("Non-specific services"), 
                  plots.g1[[2]], plots.g1[[3]], plots.g1[[5]], 
                  ncol = 2, nrow = 2, common.legend = T, legend = "none", align = "hv")
g1.b <- ggarrange(plots.g1[[6]] + ggtitle(levels(dat.plots.type$grp.cancer)[2]), plots.g1[[7]], 
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")
g1.c <- ggarrange(plots.g1[[8]] + ggtitle(levels(dat.plots.type$grp.cancer)[3]), plots.g1[[9]], plots.g1[[10]], 
                  ncol = 2, nrow = 2, common.legend = T, legend = "none", align = "hv")
g1.d <- ggarrange(plots.g1[[12]] + ggtitle(levels(dat.plots.type$grp.cancer)[4]), plots.g1[[11]], 
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")
g1.e <- ggarrange(plots.g1[[13]] + ggtitle(levels(dat.plots.type$grp.cancer)[5]), plots.g1[[14]], plots.g1[[15]], 
                  ncol = 2, nrow = 2, common.legend = T, legend = "none", align = "hv")
g1.f <- ggarrange(plots.g1[[17]] + ggtitle(levels(dat.plots.type$grp.cancer)[6]), plots.g1[[16]], 
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")

g1.type <- ggarrange(g1.a, g1.b, g1.c, g1.d, g1.e, g1.f, ncol = 1, heights = c(2,1,2,1,2,1),
          common.legend = T, legend = "top")

g1 <- ggarrange(g1.agg, g1.type, ncol = 1, heights = c(1, 6),
          common.legend = T, legend.grob = get_legend(plots.g1[[1]]), legend = "top")


ggsave(plot = g1,
       filename = "output/plots/paper prestaciones fig1 prevision.pdf",
       device = "pdf",
       dpi = "retina",
       width = 6, height = 18)
```

```{r FIG1 POR SEXO, fig.width=6, fig.height=10, message=FALSE, warning=FALSE}
g1.agg <- dat.plots.agg %>% 
  filter(mod.int == "Sex") %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019")) %>% 
  group_by(week, year, exposicion, sexo) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, sexo)%>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = mean(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb)) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  geom_point(alpha = .3, aes(color = exposicion), size = .7) +
  geom_line(aes(x = week, y = yhat, color = exposicion, group = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +
  geom_line(aes(x = week, y = cf, color = exposicion, group = exposicion, linetype = "Predicted counterfactual"), 
            size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  ylim(0, NA) +
  scale_x_discrete(breaks = seq(5, 53, 6)) +
  facet_grid(. ~ sexo) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       subtitle = "", title = "Total cancer-related health services") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 8),
        plot.title = element_text(size = 11, face = "bold"))

plots.g1 <- list()

for (i in 1:9){
  plots.g1[[i]] <- dat.plots.type.sexo %>% 
  filter(grupo.trz == levels(dat.plots.type.sexo$grupo.trz)[i]) %>% 
  mutate(exposicion = ifelse(year == 2020, "2020", "2018-2019")) %>% 
  group_by(week, year, exposicion, grupo.trz, sexo) %>% summarise(count = sum(count),
                                                 yhat = sum(yhat),
                                                 cf = sum(cf),
                                                 ub = sum(ub),
                                                 lb = sum(lb),
                                                 cf_ub = sum(cf_ub),
                                                 cf_lb = sum(cf_lb) ) %>% 
  group_by(week, exposicion, grupo.trz, sexo) %>% summarise(count = mean(count),
                                          yhat = mean(yhat),
                                          cf = mean(cf),
                                          ub = mean(ub),
                                          lb = mean(lb),
                                          cf_ub = mean(cf_ub),
                                          cf_lb = mean(cf_lb)) %>% 
  ggplot(aes(x = week, y = count, fill = exposicion))+
  #geom_point(alpha = .3, aes(color = exposicion), size = .7) +
  geom_line(aes(x = week, y = yhat, color = exposicion, group = exposicion, linetype = "Observed")) +
  geom_ribbon(aes(ymin = lb, ymax = ub, group = exposicion), alpha = .3) +
  geom_line(aes(x = week, y = cf, color = exposicion, group = exposicion, linetype = "Predicted counterfactual"), 
            size = .3, name = "bla") +
  geom_ribbon(aes(ymin = cf_lb, ymax = cf_ub, group = exposicion), alpha = .15) +
  scale_linetype_manual(values = c("Observed" = 1, "Predicted counterfactual" = 2)) +
  ylim(0, NA) +
  scale_x_discrete(breaks = seq(11, 53, 12)) +
  facet_grid(. ~ sexo) +
  theme_minimal() + 
  labs(y = "Count", x = "Week", fill = "Period", color = "Period", linetype = "Scenario",
       subtitle = levels(dat.plots.type.sexo$grupo.trz)[i], title = "") +
  geom_vline(xintercept = 11, linetype = "dotted", color = "red") + 
  theme(legend.position = "bottom",
        axis.title = element_text(size = 7),
        axis.text = element_text(size = 6),
        plot.subtitle = element_text(size = 8),
        plot.title = element_text(size = 11, face = "bold"))
}


g1.a <- ggarrange(plots.g1[[1]] + ggtitle("Non-specific services"), 
                  plots.g1[[2]], plots.g1[[3]], plots.g1[[5]], 
                  ncol = 2, nrow = 2, common.legend = T, legend = "none", align = "hv")
g1.b <- ggarrange(plots.g1[[6]] + ggtitle(levels(dat.plots.type$grp.cancer)[2]), plots.g1[[7]], 
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")
g1.c <- ggarrange(plots.g1[[8]] + ggtitle(levels(dat.plots.type$grp.cancer)[4]), plots.g1[[9]], 
                  ncol = 2, nrow = 1, common.legend = T, legend = "none", align = "hv")

g1.type <- ggarrange(g1.a, g1.b, g1.c, ncol = 1, heights = c(2,1,1),
          common.legend = T, legend = "top")

g1 <- ggarrange(g1.agg, g1.type, ncol = 1, heights = c(1, 3),
          common.legend = T, legend.grob = get_legend(plots.g1[[1]]), legend = "top")


ggsave(plot = g1,
       filename = "output/plots/paper prestaciones fig1 sexo.pdf", 
       device = "pdf",
       dpi = "retina",
       width = 6, height = 10)
```
